# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bill's Pro API",
        "description": "API documentation for Bill's Pro - Crypto and Bill Payment System",
        "contact": {
            "name": "API Support",
            "email": "support@billspro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/withdrawal/bank-accounts": {
            "get": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Get user bank accounts",
                "description": "Get all active bank accounts for the authenticated user.",
                "operationId": "12266ea8129886a0d68e0cb7cecefdb1",
                "responses": {
                    "200": {
                        "description": "Bank accounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bank accounts retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Add bank account",
                "description": "Add a new bank account for the authenticated user.",
                "operationId": "ad2f2563de6bc8e7ba5acbf76b022dfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_name",
                                    "account_number",
                                    "account_name"
                                ],
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "example": "Access Bank"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "account_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "NGN"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "NG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bank account added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bank account added successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bank account already exists"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/withdrawal/bank-accounts/{id}": {
            "put": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Update bank account",
                "description": "Update an existing bank account for the authenticated user.",
                "operationId": "5cc6ef5b5d891de49529c4913571b194",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "example": "Access Bank"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "account_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bank account updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bank account updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank account not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Delete bank account",
                "description": "Delete (deactivate) a bank account for the authenticated user.",
                "operationId": "5cbf7ce835ef5ea9b1f9f30496f99f4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bank account deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank account not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/withdrawal/bank-accounts/{id}/set-default": {
            "post": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Set default bank account",
                "description": "Set a bank account as the default account for withdrawals.",
                "operationId": "40d157c0eb3575b36434bc6cc820f37a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bank account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default bank account set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Default bank account set successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank account not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/withdrawal/fee": {
            "get": {
                "tags": [
                    "Withdrawal"
                ],
                "summary": "Get withdrawal fee",
                "description": "Get the withdrawal fee amount.",
                "operationId": "adbb310eb362d463c4f1bd96c986f292",
                "responses": {
                    "200": {
                        "description": "Withdrawal fee retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Withdrawal fee retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}